requirement.txt
mysql-connector-python==9.3.0
pillow==11.2.1
pygame==2.6.1
